put test pulse thread under the control of the clamp device
